move proxy controller to new world
