+++ po/POTFILES.in
@@ -5,6 +5,7 @@ subiquity/client/controllers/__init__.py
 subiquity/client/controllers/keyboard.py
 subiquity/client/controllers/network.py
 subiquity/client/controllers/progress.py
+subiquity/client/controllers/proxy.py
 subiquity/client/controllers/refresh.py
 subiquity/client/controllers/welcome.py
 subiquity/client/controllers/zdev.py
@@ -36,7 +37,6 @@ subiquity/controllers/filesystem.py
 subiquity/controllers/identity.py
 subiquity/controllers/__init__.py
 subiquity/controllers/mirror.py
-subiquity/controllers/proxy.py
 subiquity/controllers/reboot.py
 subiquity/controllers/snaplist.py
 subiquity/controllers/ssh.py
@@ -124,6 +124,7 @@ subiquity/server/controllers/keyboard.py
 subiquity/server/controllers/locale.py
 subiquity/server/controllers/network.py
 subiquity/server/controllers/package.py
+subiquity/server/controllers/proxy.py
 subiquity/server/controllers/refresh.py
 subiquity/server/controllers/reporting.py
 subiquity/server/controllers/userdata.py
