+++ subiquity/controllers/__init__.py
@@ -17,7 +17,6 @@ from ..controller import RepeatedController
 from .filesystem import FilesystemController
 from .identity import IdentityController
 from .mirror import MirrorController
-from .proxy import ProxyController
 from .reboot import RebootController
 from .snaplist import SnapListController
 from .ssh import SSHController
@@ -26,7 +25,6 @@ __all__ = [
     'FilesystemController',
     'IdentityController',
     'MirrorController',
-    'ProxyController',
     'RebootController',
     'RepeatedController',
     'SnapListController',
